[#48] Fix useSearchParams() Suspense error
