+++ frontend/app/(routes)/(auth)/activate/page.tsx
@@ -1,48 +1,12 @@
 'use client'
 
-import api from '@/app/_api/api'
-import Link from 'next/link'
-import { useSearchParams } from 'next/navigation'
-import { useRouter } from 'next/navigation'
-import { useEffect, useState } from 'react'
+import Activate from '@/components/activate'
+import { Suspense } from 'react'
 
 export default function Verify() {
-  const [verificationError, setVerificationError] = useState<boolean>(false)
-  const [loading, setLoading] = useState(false)
-
-  const searchParams = useSearchParams()
-  const router = useRouter()
-  const token = searchParams.get('token')
-  const uid = searchParams.get('uid')
-
-  useEffect(() => {
-    setLoading(true)
-    let fetchApi = async () => {
-      try {
-        const { status } = await api.post(`/auth/users/activation`, {
-          token,
-          uid,
-        })
-        if (status >= 200 && status < 300) {
-          router.push('/')
-        }
-      } catch (error: any) {
-        setLoading(false)
-        setVerificationError(true)
-      }
-    }
-    fetchApi()
-  }, [router, token, uid])
-
   return (
-    <div>
-      {loading && <div>Activating account...</div>}
-      {verificationError && (
-        <div style={{ color: 'red' }}>
-          The link is invalid and has likely expired, please try&nbsp;
-          <Link href='/register'>resending the email.</Link>
-        </div>
-      )}
-    </div>
+    <Suspense fallback={<div>Loading page...</div>}>
+      <Activate />
+    </Suspense>
   )
 }
